# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /swarm/simulate:
    post:
      operationId: simulateSwarm
      tags: [swarm]
      description: Synthetic scenarios are public. AI Village requests require verified Clerk authentication plus explicit operator enablement and a reader allowlist; disabled by default.
      summary: Replay a synthetic scenario (enforce) or an AI Village window (report-only) through ASP and Sentinel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunInput"
      responses:
        "401": { description: AI Village authentication required or expired }
        "403": { description: Private AI Village access not approved or untrusted mutation origin }
        "503": { description: Authentication unavailable or store unavailable }
        "200":
          description: Recorded simulation and forensic report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwarmRun"
...
  /swarm/sessions:
    get:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: listSwarmSessions
      tags: [swarm]
      summary: Live sessions, newest first
      responses:
        "401": { description: Authentication required or expired }
        "503": { description: Authentication unavailable or not configured }
        "200":
          description: Session summaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SessionSummary" }
    post:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: createSwarmSession
      tags: [swarm]
      summary: Start a live session (stateful gateway and Sentinel, enforce mode)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SessionInput" }
      responses:
        "401": { description: Authentication required or expired }
        "403": { description: Untrusted cookie mutation origin }
        "429": { description: Session capacity reached }
        "503": { description: Authentication unavailable or not configured }
        "201":
          description: Session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionCreated" }
  /swarm/sessions/{sessionId}/stream:
    get:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: streamSwarmSession
      tags: [swarm]
      summary: "Owner-only SSE; auth-expired closes the stream at JWT expiry. Reconnect with a fresh snapshot."
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
        - { name: after, in: query, required: false, schema: { type: integer, minimum: 0 } }
      responses:
        "401": { description: Authentication required or expired }
        "503": { description: Authentication unavailable or not configured }
        "200":
          description: "text/event-stream; each data line is a SessionStreamMessage"
          content:
            text/event-stream:
              schema: { type: string }
        "404": { description: Unknown or expired session }
  /swarm/demo/injection:
    post:
      operationId: launchInjectionDemo
      tags: [swarm]
      summary: Start a PUBLIC SYNTHETIC demo in isolated storage; no caller-supplied traces or private session access
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DemoInput" }
      responses:
        "202":
          description: Demo started
          content:
            application/json:
              schema:
                type: object
                required: [sessionId]
                properties:
                  sessionId: { type: string }
  /swarm/sessions/{sessionId}/sandbox:
    post:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: configureSwarmSandbox
      tags: [swarm]
      summary: Initialize server-owned fictional fixtures once before agent actions
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SandboxInput" }
      responses:
        "201":
          description: Disposable sandbox initialized; no filesystem paths are returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SandboxSnapshot" }
        "401": { description: Authentication required or expired }
        "403": { description: Untrusted cookie mutation origin }
        "404": { description: Unknown or expired session }
        "409": { description: Wrong policy or sandbox already initialized or actions already recorded }
        "422": { description: Invalid fixture mode or extra fields }
        "503": { description: Authentication unavailable or not configured }
    get:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: getSwarmSandbox
      tags: [swarm]
      summary: Read owner-private sandbox execution evidence
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Actual fixture ledger, board and successful tool-body list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SandboxSnapshot" }
        "401": { description: Authentication required or expired }
        "404": { description: Unknown or expired session }
        "409": { description: Sandbox is not initialized }
        "503": { description: Authentication unavailable or not configured }
  /swarm/sessions/{sessionId}/execute:
    post:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: executeSwarmTool
      tags: [swarm]
      summary: Evaluate and execute a registered sandbox tool under one session lock
      description: >-
        Enforcement metadata is server-owned. A denied decision never reaches a tool body.
        A required idempotencyKey is scoped to this disposable session and bound to the full
        execution payload (actor, lineage, tool and exact arguments). An exact replay returns
        the original outcome, including refusals and failures, without reevaluation or dispatch.
        A changed payload with the same key is refused. New keys always receive fresh ASP checks.
        Authentication and ownership are checked on every request, including replays.
        Keys and outcomes last for the session lifetime only; no recovery across process restarts
        is promised. No caller-supplied decision, path, trust label or credential is accepted.
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExecutionInput" }
      responses:
        "200":
          description: Gateway decision plus actual result or refusal; allowed is permission, not proof of tool success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolOutcome" }
        "401": { description: Authentication required or expired }
        "403": { description: Untrusted cookie mutation origin }
        "404": { description: Unknown or expired session }
        "409": { description: "Sandbox unavailable, idempotency key payload conflict or interrupted execution with unavailable outcome" }
        "422": { description: "Unknown tool, invalid arguments or caller-supplied enforcement metadata" }
        "503": { description: Authentication unavailable or not configured }
  /swarm/sessions/{sessionId}/completion:
    post:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: completeSwarmTool
      tags: [swarm]
      summary: Record caller-reported completion of an allowed permission-only tool
      description: >-
        The owner must supply the session-bound completionToken returned by evaluate
        and the exact event, actor and span. Only one terminal receipt is accepted.
        Refused calls, non-tools, sandbox executions, and duplicate or contradictory
        receipts cannot be updated. No results or error text are accepted.
        This is caller-reported evidence, not engine observation or proof of side effects.
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompletionReceipt" }
      responses:
        "200":
          description: Updated existing recorder row
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trace" }
        "401": { description: Authentication required or expired }
        "403": { description: Untrusted cookie mutation origin }
        "404": { description: Unknown session or event }
        "409": { description: "Ineligible, mismatched, duplicate or contradictory receipt" }
        "422": { description: Invalid receipt }
        "503": { description: Authentication unavailable or not configured }
  /swarm/sessions/{sessionId}/evaluate:
    post:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: evaluateSwarmCall
      tags: [swarm]
      summary: Decide one proposed agent action before it executes
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CallInput" }
      responses:
        "401": { description: Authentication required or expired }
        "403": { description: Untrusted cookie mutation origin }
        "503": { description: Authentication unavailable or not configured }
        "200":
          description: Decision; allowed=false means the agent must not execute the call
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CallDecision" }
        "404": { description: Unknown or expired session }
  /swarm/sessions/{sessionId}:
    get:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: getSwarmSession
      tags: [swarm]
      summary: Snapshot of a live session as a run record
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        "401": { description: Authentication required or expired }
        "503": { description: Authentication unavailable or not configured }
        "200":
          description: Live run record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SwarmRun" }
        "404": { description: Unknown or expired session }
    delete:
      security: [{ ClerkCookie: [] }, { ClerkBearer: [] }]
      operationId: deleteSwarmSession
      tags: [swarm]
      summary: End a live session
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        "401": { description: Authentication required or expired }
        "403": { description: Untrusted cookie mutation origin }
        "503": { description: Authentication unavailable or not configured }
        "204": { description: Deleted }
        "404": { description: Unknown or expired session }
...
  /swarm/demo/sessions/{sessionId}:
    get:
      operationId: getSwarmDemoSession
      tags: [swarm]
      summary: Public synthetic snapshot and report only
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Synthetic demonstration record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SwarmRun" }
        "404": { description: Unknown or expired demo; includes private session IDs }
  /swarm/demo/sessions/{sessionId}/stream:
    get:
      operationId: streamSwarmDemoSession
      tags: [swarm]
      summary: Public synthetic SSE only
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
        - { name: after, in: query, required: false, schema: { type: integer, minimum: 0 } }
      responses:
        "200":
          description: Synthetic SessionStreamMessage data lines
          content:
            text/event-stream:
              schema: { type: string }
        "404": { description: Unknown or expired demo; includes private session IDs }
...
components:
  securitySchemes:
    ClerkCookie:
      type: apiKey
      in: cookie
      name: __session
      description: Verified Clerk session JWT; browser requests use cookies. Cookie mutations require an approved Origin.
    ClerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Verified Clerk session JWT for remote SDKs; never an arbitrary identity header. All private routes are owner-only and return 404 for non-owners.
  schemas:
    RunInput:
      type: object
      required: [scenario, feedbackEnabled]
      properties:
        scenario: { type: string, enum: [normal, attack, injection, runaway, ai-village] }
        feedbackEnabled: { type: boolean }
        maxDepth: { type: integer, minimum: 1, maximum: 10, description: Defaults to the scenario policy }
        semanticLimit: { type: integer, minimum: 2, maximum: 20, description: Defaults to the scenario policy }
        writeLimit: { type: integer, minimum: 1, maximum: 60, description: Defaults to the scenario policy }
        episodeId: { type: string, description: Indexed AI Village episode to replay }
        start: { type: string, description: UTC start of an AI Village window (with end; max 24 hours) }
        end: { type: string }
        maxEvents: { type: integer, minimum: 10, maximum: 20000, default: 5000 }
...
    SwarmRun:
      type: object
      required: [id, scenario, provenance, events, alerts, policies, report, rootAgent, startedAt]
      properties:
        id: { type: string }
        scenario: { type: string }
        provenance: { type: string }
        executionEvidence: { type: boolean, description: "Owner-private Live separates authorization, engine-observed outcomes and caller-reported receipts. Absent for synthetic demos and historical replays." }
        completionGraceSeconds: { type: integer, minimum: 1, maximum: 86400, description: "Owner-private only. Seconds since server admission before an unreported external completion is overdue." }
        completionWarnings: { type: array, items: { $ref: "#/components/schemas/CompletionWarning" }, description: "Owner-private visibility only. Missing receipts do not establish body entry, failure, rollback or retry safety; separate from detections and enforcement." }
        events: { type: array, items: { $ref: "#/components/schemas/Trace" } }
        alerts: { type: array, items: { $ref: "#/components/schemas/SwarmAlert" } }
        policies: { type: array, items: { $ref: "#/components/schemas/PolicyChange" } }
        report: { type: string }
        rootAgent: { type: string }
        startedAt: { type: string }
        source: { type: string, enum: [synthetic, ai-village, live] }
        mode: { type: string, enum: [enforce, report-only] }
        policyName: { type: string }
        policy: { type: object, additionalProperties: true }
        window: { oneOf: [{ $ref: "#/components/schemas/ReplayWindow" }, { type: "null" }] }
        summary: { $ref: "#/components/schemas/RunSummary" }
        edges:
          type: array
          description: Executed interactions collapsed to one edge per directed pair, heaviest first
          items: { $ref: "#/components/schemas/WeightedEdge" }
...
    SandboxInput:
      type: object
      additionalProperties: false
      properties:
        mode: { type: string, enum: [normal, adversarial], default: normal }
    ExecutionInput:
      type: object
      additionalProperties: false
      required: [idempotencyKey, agentId, spanId, tool]
      properties:
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[A-Za-z0-9._:-]+$'
          description: Unique key for one logical execution. Reuse it only with the unchanged payload to recover a lost response.
        agentId: { type: string, minLength: 1, maxLength: 120 }
        spanId: { type: string, minLength: 1, maxLength: 200 }
        parentId: { type: string, maxLength: 120, default: "" }
        parentSpanId: { type: string, maxLength: 200, default: "" }
        tool: { type: string, enum: [read_invoice, read_ledger, transfer_funds, post_status] }
        arguments:
          type: object
          additionalProperties: true
          description: "Validated against the selected tool: reads accept {}; transfer_funds requires positive integer amount and recipient approved-supplier or sandbox-attacker; post_status requires nonblank text of 1–500 characters. No other keys."
...
    SandboxStatus:
      type: object
      required: [text]
      properties:
        text: { type: string }
    SandboxSnapshot:
      type: object
      required: [sandboxOnly, mode, initialCredits, ledger, board, actualToolBodiesExecuted]
      properties:
        sandboxOnly: { type: boolean, const: true }
        mode: { type: string, enum: [normal, adversarial] }
        initialCredits: { type: integer }
        ledger: { $ref: "#/components/schemas/SandboxLedger" }
        board: { type: array, items: { $ref: "#/components/schemas/SandboxStatus" } }
        actualToolBodiesExecuted: { type: array, items: { type: string } }
...
    SessionInput:
      type: object
      properties:
        policy: { type: string, enum: [mock, ai-village, agents], default: mock }
        feedbackEnabled: { type: boolean, default: true }
        root: { type: string, default: orchestrator }
        label: { type: string, maxLength: 200, default: "" }
        completionGraceSeconds: { type: integer, minimum: 1, maximum: 86400, default: 60, description: "Server-admission grace period for missing external completion receipts. Visibility only; not a tool timeout or retry deadline." }
...
    SessionSummary:
      type: object
      required: [sessionId, created, policyName, root, label, events, alerts]
      properties:
        sessionId: { type: string }
        created: { type: string }
        policyName: { type: string }
        root: { type: string }
        label: { type: string }
        events: { type: integer }
        alerts: { type: integer }
...
    DemoInput:
      type: object
      additionalProperties: false
      properties:
        feedbackEnabled: { type: boolean, default: true }
        pause: { type: number, minimum: 0, maximum: 5, default: 1.2 }
    SessionCreated:
      type: object
      required: [sessionId, policyName, root, feedbackEnabled, policy]
      properties:
        sessionId: { type: string }
        policyName: { type: string }
        root: { type: string }
        feedbackEnabled: { type: boolean }
        completionGraceSeconds: { type: integer, minimum: 1, maximum: 86400 }
        policy: { type: object, additionalProperties: true }
...